Refactor flight report infrastructure
