--- orcestra_book/operation.md
@@ -1,2 +1,0 @@
-<!-- This file will be updated automatically -->
-# Operation
